[chore] Bit more refactoring of settings panel
